requirements.txt
asgiref==3.6.0
autopep8==2.0.2
beautifulsoup4==4.12.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
defusedxml==0.7.1
Django==4.1.7
django-bs4==0.0.4
django-cors-headers==3.14.0
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gcloud==0.18.3
googleapis-common-protos==1.59.0
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
jws==0.1.3
MarkupSafe==2.1.2
oauth2client==4.1.3
oauthlib==3.2.2
Pillow==9.4.0
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.17
PyJWT==2.6.0
pyparsing==3.0.9
Pyrebase==3.0.27
Pyrebase4==4.6.0
python-jwt==2.0.1
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.0
soupsieve==2.4
sqlparse==0.4.3
stripe==5.2.0
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.15
